+++ CHANGES.md
@@ -1,6 +1,9 @@
 Change Log
 ==========
 
+### Next Release
+* Fixed `combinePrimitives` stage and re-added it to the pipeline. [#108](https://github.com/AnalyticalGraphicsInc/gltf-pipeline/issues/108)
+
 ### 0.1.0-alpha5 - 2016-11-02
 
 * Added `MergeDuplicateProperties` for stages merging duplicate glTF properties, like materials and shaders. [#152](https://github.com/AnalyticalGraphicsInc/gltf-pipeline/pull/152)